src/cellsepi/pyproject.toml
[tool.flet.flutter.pubspec.dependency_overrides]
webview_flutter_android = "4.10.1"

[project]
dependencies = [
    "flet-extended-interactive-viewer == 0.2.0",
]